add lesson5 of module6 - add blog`s pages
