+++ blogs/blog.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="ru">
+
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<link href="../normalize.css" rel="stylesheet" type="text/css">
+	<link rel="preconnect" href="https://fonts.googleapis.com">
+	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+	<link href="https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap" rel="stylesheet">
+	<link type="text/css" rel="stylesheet" href="style_blog.css">
+	<script type="module" src="../script/script.js"></script>
+	<title>Blog</title>
+</head>
+
+<body>
+	<div class="page">
+		<div class="wrapper">
+			<div class="blog">
+				<!--карточки товаров-->
+			</div>
+
+			<div class="pagination">
+				<div class="pagination__pages">
+					<button class="pagination__pages-arrow-left"></button>
+					<div class="pagination__pages-numbers">
+						<!--пагинация-->
+					</div>
+					<button class="pagination__pages-arrow-right"></button>
+				</div>
+			</div>
+
+		</div>
+	</div>
+</body>
+
+</html>
