--- order/global_style.css
@@ -1,84 +1,0 @@
-html {
-	position: relative;
-	box-sizing: border-box;
-	overflow: scroll;
-	overflow-x: scroll;
-	scroll-behavior: smooth;
-}
-
-*,
-*::before,
-*::after {
-	box-sizing: inherit;
-}
-
-body {
-	font-family: 'Inter', sans-serif;
-	font-size: 16px;
-	min-width: 320px;
-	min-height: 100vh;
-}
-
-img,
-svg {
-	max-width: 100%;
-	max-height: 100%;
-	height: auto;
-}
-
-button {
-	padding: 0;
-	cursor: pointer;
-}
-
-a {
-	text-decoration: none;
-	color: inherit;
-}
-
-ul {
-	list-style: none;
-	padding: 0;
-	margin: 0;
-}
-
-h1,
-h2,
-h3,
-p,
-figure,
-fieldset {
-	margin: 0;
-}
-
-iframe {
-	border: none;
-}
-
-/*
- Доступно скрываю элементы
- */
-
-.visually-hidden {
-	position: fixed;
-	transform: scale(0);
-}
-
-/*
- Пример контейнера
- 
- Если контейнер 1180px по макету,
- то для max-width прибавляем паддинги 30px*2, 
- в итоге 1240px
- Чтобы контент не прижимался к краям экрана
- */
-
-.container {
-	max-width: 1240px;
-	padding: 0 30px;
-	margin: 0 auto;
-}
-
-blockquote {
-	margin: 0;
-}
