+++ order/index_cms.html
@@ -8,7 +8,7 @@
 	<link rel="preconnect" href="https://fonts.googleapis.com">
 	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 	<link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" rel="stylesheet">
-	<link href="./normalize.css" rel="stylesheet" type="text/css">
+	<link href="../normalize.css" rel="stylesheet" type="text/css">
 	<link href="./style_order.css" rel="stylesheet" type="text/css">
 	<title>Перечень товаров</title>
 </head>
