+++ order/style_order.css
@@ -1,4 +1,4 @@
-@import url("./global_style.css");
+@import url("../global_styles.css");
 
 @import url("./styles_order/overlay.css");
